requirements.txt
streamlit
requests
faiss-cpu
transformers
sentencepiece
protobuf
ftfy
usearch
huggingface_hub
tqdm
